extract iwTree.cpp
