+++ callbacks.h
@@ -3,14 +3,16 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
+#pragma once
+
+#include "iwTree.h"
+
 // NOTE: negative callbackParams are reserved: -1 = callback is called first time, -2 = used by gameloop for registered
 // callbacks (callbacks that will additionally execute WITHIN the gameloop)
 
 // NOTE: don't forget that if the map quits (Param: MAP_QUIT), there are many windows that have to be closed.
 // This happens for example if a new Map will be loaded or user goes to main menu. So if you add a new window, don't
 // forget to add it to this "close lists" if it's necessary (also in the file CMap.cpp, function setMouseData(Button)).
-
-#pragma once
 
 namespace callback {
 // PleaseWait creates a small window (not moveable, not resizeable, not minimizable, not closeable) with the String
@@ -26,7 +28,6 @@ void EditorLoadMenu(int Param);
 void EditorSaveMenu(int Param);
 void EditorQuitMenu(int Param);
 void EditorTextureMenu(int Param);
-void EditorTreeMenu(int Param);
 void EditorResourceMenu(int Param);
 void EditorLandscapeMenu(int Param);
 void EditorAnimalMenu(int Param);
