+++ iwTree.h
@@ -1,0 +1,10 @@
+// Copyright (C) 2009 - 2021 Marc Vester (XaserLE)
+// Copyright (C) 2009 - 2021 Settlers Freaks <sf-team at siedler25.org>
+//
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#pragma once
+
+namespace callback {
+void EditorTreeMenu(int Param);
+} // namespace callback
